requirements.txt
coqui-tts==0.24.2
silero_vad==5.1.2
numpy==1.25.2 # Adjusted for compatibility
torch==2.4.1
torchaudio==2.4.1
ollama==0.3.3
faster_whisper==1.0.3
langchain_anthropic==0.2.3
langchain_community==0.3.2
langchain_google_genai==2.0.1
langchain_groq==0.2.0
langchain_huggingface==0.1.0
langchain_openai==0.2.2
langchain-ollama==0.2.0
PyAudio==0.2.14
pytz==2024.2
pyttsx3==2.98
